Add hack which makes the backend layer 2 display the frontend's output.
Leaving the define HACK_BACKEND_LAYER2_TO_FRONTEND enabled, this driver will
use layer 2 of the backend to output of the frontend. This allows the driver
to input the decoded MPEG-2 output from the cedrus driver (tiled YUV420) and
show the detiled ARGB8888 output to be displayed on whatever connector is
connected to the backend.

Signed-off-by: Thomas van Kleef <linux-dev@vitsch.nl>